Decreased number of edges when using `allowSubmatch`
Previously, `allowSubmatch` increases the number of edges exponentially.
That makes code size too big and compile time too long!

It now uses indirect transition when a matching fails and go back to
the start state, instead of directly jump to next to the start state.
That can decrease the number of edges drastically, but requires one
more jump as overhead when a matching fails.